+++ test/files/psql/schema.sql
@@ -1,19 +1,25 @@
 -- Don't forget to maintain order here, foreign keys!
 drop table if exists video_tags;
 drop table if exists tags;
+drop view if exists user_videos;
 drop table if exists videos;
 drop table if exists sponsors;
 drop table if exists users;
-drop table if exists type_monsters;
-drop view if exists user_videos;
-drop view if exists type_monsters_v;
 drop materialized view if exists type_monsters_mv;
+drop view if exists type_monsters_v;
+drop table if exists type_monsters;
+drop table if exists test_index_expressions;
+drop table if exists foo_bar;
+drop table if exists foo_baz;
+drop table if exists foo_qux;
+drop table if exists bar_baz;
+drop table if exists bar_qux;
 
 drop type if exists unicode_enum;
 create type unicode_enum as enum ('hello', 'привет', 'こんにちは', '안녕하세요', 'hello_with_underscore');
 
 drop domain if exists uint3;
-create domain uint3 as numeric check(value >= 0 and value < power(2::numeric, 3::numeric));
+create domain uint3 as numeric;
 
 create table users (
 	id serial primary key not null,
@@ -245,13 +251,14 @@ create table type_monsters (
 	polygon_null  polygon NULL,
 	polygon_nnull polygon NOT NULL,
 
-	tsquery_null   tsquery NULL,
-	tsquery_nnull  tsquery NOT NULL,
 	tsvector_null  tsvector NULL,
 	tsvector_nnull tsvector NOT NULL,
 
 	txid_null  txid_snapshot NULL,
 	txid_nnull txid_snapshot NOT NULL,
+
+	pg_snapshot_null  pg_snapshot NULL,
+	pg_snapshot_nnull pg_snapshot NOT NULL,
 
 	xml_null  xml NULL,
 	xml_nnull xml NOT NULL,
@@ -298,7 +305,7 @@ create table type_monsters (
 
     base text null,
 
-    generated_nnull text NOT NULL GENERATED ALWAYS AS (UPPER(base)) STORED,
+    generated_nnull text NOT NULL GENERATED ALWAYS AS (UPPER(string_two)) STORED,
     generated_null text NULL GENERATED ALWAYS AS (UPPER(base)) STORED
 );